generate: Change the type of DeepCopyInto() back
According to the comment in
k8s.io/gengoi/examples/deepcopy-gen/generators/deepcopy.go
deepCopyIntoMethod(), the type of the parameter of DeepCopyInto()
should be "t *T" and not "t **T". So use that in the expression.go.

Doing this causes compile error in DeepCopyInto() in generated
zz_generated.deepcopy.go file:

pkg/apis/resmgr/v1alpha1/zz_generated.deepcopy.go:107:26: cannot \
  use &(*out)[i] (value of type **resmgr.Expression) as \
  type *resmgr.Expression in argument to (*in)[i].DeepCopyInto

Assume here that the zz_generated.deepcopy.go is wrong and buggy.
So we are manually fixing the generated file here.

Signed-off-by: Jukka Rissanen <jukka.rissanen@linux.intel.com>