e2e: Add support for Github self hosted action runner
The scripts in this directory will:

* Provision a Docker based container to be used by the e2e tests
* Run a self-hosted action runner which will setup a VM for the
  tests, and then discard the VM after the tests have finished.

You need first to provision the base image that contains most of
the software that is needed by the e2e tests. See the README.md
for instructions how to create the setup.
The container will then run, and execute Github self hosted runner
script which will then create a VM where the e2e tests are executed.
The self hosted runner is triggered to run the e2e tests from Github
workflow actions yaml file.

Signed-off-by: Jukka Rissanen <jukka.rissanen@linux.intel.com>