e2e: Add a script that installs cilium
Use a special script to install cilium in order to make sure
that it is running when the e2e tests are run. The script will
retry the installation couple of times before giving up.
It is rather pointless to run the tests if the cilium
installation failed as the tests will most likely fail.

Signed-off-by: Jukka Rissanen <jukka.rissanen@linux.intel.com>