cache, topology: fix hint generation bugs.
When inserting containers into the cache, make sure host
and container paths for mounts are passed in the correct
order to the hint generator function.

For each injected device, expect a corresponding cgroup
access rule and use it to determine if the device is in
read-only mode.

If there is a sysRoot given, use it when looking for a
sysfs device based on path, or type/major/minor combo.

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>