topology-aware: pick reserved pool by CPU type alone.
Do not second guess the already determined CPU type for
containers in the kube-system namespace by checking the
namespace again. Short circuit assigning reserved-pool
containers to the root based on the CPU type alone.

Note that this also allows containers to be opted out
from the reserved pool using an explicit
  prefer-reserved-cpus.resource-policy.nri.io/pod: "false"
annotation.

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>