policy: Update container allocations at startup
When policy is started, it reads the saved container allocations
from cache file. This can cause problems if we change the resource
allocation policy as the new policy would use the previous resource
allocations and not generate allocations based on its own policy.
Solve this by forcing re-allocation of resources at startup.

Signed-off-by: Jukka Rissanen <jukka.rissanen@linux.intel.com>