e2e: Get cache from HTTP instead of file
The e2e tests want to validate things after the test and for that,
they need access to the cache state. So now that the cache file is
removed, the cache state needs to be fetched directly from NRI
plugin via a HTTP request.
Default address used by e2e tests is http://localhost:8891/cache-state
The cache fetching is disabled by default and only enabled for
e2e tests.

Signed-off-by: Jukka Rissanen <jukka.rissanen@linux.intel.com>