Makefile: better plugin binary dependencies.
We want each plugin binary to depend on the go sources of their
(in-tree) go dependecies and the plugin 'main' package sources.
`go list` does not list 'main', so we need to inject it manually
to the generated dep list, which is what the first for-echo loop
does. However, we need to make sure that our dep-filtering grep
keeps both in-tree module dependencies and the main package for
each plugin. This commit fixes the latter.

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>