cache: compile hint-ignore regexp only once.
There is no need to compile the hint-ignoring static regexp
every time we use it. As an additional bonus if MustCompile
ever fails with a panic due to someone incorrectly updating
the regexp, it will do so already during startup instead of
when we try to generate hints for the first time.

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>