cache: add functions for dealing with resource updates.
Add container functions for storing and retrieving resource
updates to and from the cache. The updates are converted to
CRI resource requirement before storing. The same format is
used used to store the original requirements, so these can
now be used interchangeably.

The storing function returns whether an update really changed
any any of the resources. This is used to filter out multiple
updates with identical resources, which we do get.

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>