instrumentation: use a more appropriate propagator.
Use a correct propagator for ttRPC (and gRPC) span propagation.
Without this, ttRPC (or gRPC) span information are not properly
transmitted over the IPC. As a side effect, spans generated at
the RPC server end of the IPC do not get properly nested under
the related spans generated on the originator side.

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>