helm: fix the configmap naming in helm charts
Our agent expects the configmap to be named as
`nri-resource-policy-config.default`, whereas in the helm chart,
we have been using different names, such as
`nri-resource-policy-topology-aware-config`. In order for the
configmap configuration data to take effect, we must rename it
to match the agent's expectation. Additionally, there are no
scenarios where two plugins can run on the same node with two
different configmap names. Therefore, we can have a consistent
and standard name for the configmaps of different policy.

Signed-off-by: Feruzjon Muyassarov <feruzjon.muyassarov@intel.com>