Automate Helm charts releasing process via GitHub action
Add two GitHub actions to automate Helm charts releasing process.
First action would be responsible for packaging the charts and
pushing them to the releases and the second action is responsible
for updating the index.yaml whenever there is a new release
available.

In other words, one GitHub action to package helm charts and upload
those charts including the source code to GitHub releases and create
a draft release. This job gets triggered when a new tag is pushed.

Second GitHub action - to officially release the charts once
official release is published. When repo maintainers, publish the
draft release this action gets triggered in updates the index.yaml
in gh-pages branch with the corresponding URLs of the chart packages.

Signed-off-by: Feruzjon Muyassarov <feruzjon.muyassarov@intel.com>