e2e: rename nri-resource-policy specific balloons tests
The goal of this patch is to use the same e2e test numbers in
nri-resource-policy and cri-resource-manager e2e tests for the
balloons policy. This makes it easier to keep e2e tests in sync when
porting policy features between these projects. Now test[01]* are
common e2e tests for both projects, and test2* are nri-policy-specific
e2e tests.

Signed-off-by: Antti Kervinen <antti.kervinen@intel.com>