images: fix the tagging docker images build from new tags
When docker image workflow is triggered based on the new tag push
which usually includes v as prefix, the workflow drops the v prefix.
As such, when v0.2.0 tag is pushed to GitHub, we end up with 0.2.0
tags in the docker images. This commit fixes the workflow to keep
the v prefix.

Signed-off-by: Feruzjon Muyassarov <feruzjon.muyassarov@intel.com>