config-manager: add kludge/workaround for CRI-O.
It looks like in the case of CRI-O we need to give it some time
after we have been started up but before we kick it in the head
to restart it over D-Bus. Otherwise it will always report a 255
(-1) exit status for us. Since we run as an init-container, a
non-zero exit status would prevent other containers in our pod
from ever starting up.

It would be good to try and find out what is the exact reason
why this behavior is exhibited by CRI-O but not containerd, and
if it could be fixed in CRI-O. Until then... this.

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>