resmgr: handle UpdateContainer requests better.
Try doing a better job of detecting and handling redundant
UpdateContainer requests which do not really change any of
a container's parameters. Such requests are typically sent
by an active CPU Manager with the static policy enabled in
kubelet.

Before evaluating an update request for redundancy fill in
any missing values from the container's current resources.
Then if the request looks redundant short-circuit handling
it by sending a reply with the current resources. Otherwise
pass it on to the policy.

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>