e2e: fix balloons test CPU frequencies
Use kHZ instead of MHz in minFreq and maxFreq as required in the real
interface. Using wrong units in the tests did not fail because
configuring CPU frequencies with any values does not really work in
our test VMs. However, using wrong units in tests is misleading for
anyone who tries to use similar configuration on bare metal.

Signed-off-by: Antti Kervinen <antti.kervinen@intel.com>