github: don't trigger docs update on non-release tags
When triggered from a tag, our gh-pages update script considers this as
a new release. This means potentially (if it's the first tag in the vX.Y
series) creating a new vX.Y/ subdir in the docs, linking that as the
latest stable release etc. We definitely don't want this to happen for
devel or rc or such tags. Modifying the workflow file to only act on
well-formed release tags (like v1.2.3) is the most straightforward way
to fix this.

Signed-off-by: Markus Lehtonen <markus.lehtonen@intel.com>
(cherry picked from commit b3dd0c8c884399895f5b01d7bf7d1f53087d9832)