.github: fix release check for Helm Chart packaging.
Setting an evironment variable and using that same variable to
trigger conditional jobs with 'if:' does not work as expected.
Use a dedicated job to decide if we are dealing with a release
tag and propagate that decision to the output of the job. Then
use the output to trigger either one of a dependent release or
unstable job.

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>