.github: fix 'v'-tag-is-release logic.
Now that we're more selective with our triggering tags, it is
safe to go with the simple 'a tag is always a release' logic.
However, evaluating the condition once then passing the result
on in the environment to alternative jobs within the workflow
does not work as expected. Therefore, evaluate the condition
directly whenever we need the result.

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>