e2e: use sorted github releases to get latest release.
Our method of using the github.com/$repo/releases/latest to
query and determine the latest release of containerd, CRI-O
or kubernetes is flawed. The semantics of that 'latest' is
calendar time of release, so it will pick wrong versions if
a new tagged maintenance release is the newest by calendar
time.

Ideally we would use the official github API to list all
tagged releases not marked as a prerelease, and pick the
latest one by tag. However, without authenticating that
API is severly rate limited and fails often from behind
busy proxies. So instead, fetch pages of releases, parse
them for release URLs, sort them by version and pick the
latest.

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>