e2e: version-sort github releases to resolve latest.
Our method of querying github.com/$repo/releases/latest
to determine the latest release of containerd, CRI-O or
kubernetes is unsafe. We don't control which version is
tagged the 'latest', as it is up to the repo maintainer.
Since mistakes happen, we can end up picking and trying
to use a too old version. So instead, list releases and
version-sort them to find the latest one.

Ideally we would use the official github API to list all
tagged releases not marked as a prerelease, and pick the
latest one by tag. However, without authenticating that
API is severly rate limited and fails often from behind
busy proxies. So instead, fetch pages of releases, parse
them for release URLs, sort them by version and pick the
latest.

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>