cpuallocator: don't filter based on single CoreKind.
If we have only one type of core present (for instance E-core)
and the caller indicates a priority preference which corresponds
to an unavailable core type (for instance high-prio/P-core), we
don't currently attempt clustered allocation at all, since the
initial CoreKind based filtering filters out all CPU clusters as
unsuitable.

This does not make sense so relax this and only do core type
filtering when we have multiple of them (when running on hybrid
core architectures).

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>