cpuallocator: discover cache groups.
Discover groups of CPUs sharing a cache. Ignore groups which
happen to be the same as CPUs grouped by some other topology
criteria (package, die, etc.). Assume that all CPUs within a
group are from the same package and die, and share the same
closest NUMA node. Verify this assumption during discovery,
aborting with a panic if it is not true.

Use the last useful level of cache for clustering, IOW try
finding the last cache which provides non-trivial clustering
and use it grouping. Note that the current implementation is
somewhat simplistic. It expects all CPUs to provide identical
cache grouping and therefore picks a single cache level. This
is usually fine, but it might result in suboptimal clustering
with hybrid cores. If necessary we can remove this limitation
in the future.

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>