balloons: add support for explicit CPUs
This commit introduces a new field in the balloons configuration CR,
enabling users to specify their CPU preferences for workloads.
When a user explicitly requests a CPU set, the policy will attempt
to allocate CPUs from the specified set. If the request cannot be
satisfied, the policy will fall back to other available CPUs.

If a user adds `preferIsolCpus` in addition to specifying preferred
CPUs, the policy will prioritize allocating isolated CPUs over the
preferred CPUs set, in case these sets are different.

Signed-off-by: Feruzjon Muyassarov <feruzjon.muyassarov@intel.com>