balloons: add support for picking efficient/performance cores
This commit introduces a new field in the balloons configuration CR,
enabling users to specify their efficient/performance core preferences
for their workloads. When a user explicitly requests efficient or
performance cores, the policy will attempt to allocate according
to the user preferences. If the request cannot be satisfied, the
policy will fall back to other available cores  on the host.

If a user adds `preferIsolCpus` in addition to specifying coreTypes,
the policy will prioritize allocating isolated CPUs cores over the
preferred efficient/performance cores, in case these sets are different.

Signed-off-by: Feruzjon Muyassarov <feruzjon.muyassarov@intel.com>