topology-aware: rework building the topology pool tree.
Rework and simplify how we build the tree of pools based on HW
topology. Build the tree by walking the topology from socket to
dies to NUMA nodes closest to dies. Assign CPU and memory with
CPU locality in tandem as we build the tree. Assign NUMA nodes
without CPU locality to any pool where at least one of their
closest neighbor with CPU locality is already assigned. In the
root pool, we assign all NUMA nodes with attached memory.

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>