e2e: drop old shims from custom-build containerd
Recent versions of containerd do not build dropped shims. It is
realistic to assume that we are building and testing with most recent
containerd versions rather than old ones, so let's support that case
in our default e2e workflow, too.

This change affects only testing with containerd_src set to a local
containerd build.

Signed-off-by: Antti Kervinen <antti.kervinen@intel.com>