config,crds: add RDT class configuration.
Add RDT class and partition configuration. If given and the
corresponding control is enabled, the configuration is used
to set up RDT CLOSes using goresctrl.

The configuration is almost identical to the goresctrl RDT
configuration, but does not support the shorthand notation
for unified caches which would be problematic for a custom
resource data type.

IOW, one cannot simply say

```
...
  l[23]Allocation:
    all: 75%
...
```

but instead needs to explicitly write this as

```
...
  l[23]Allocation:
    all:
      unified: 75%
...
```

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>