topology-aware: take exclusive CPUs by podresapi hint(s) first.
If a container asks for at least as many exsclusive CPUs as it
has pod resource API hints, try allocating CPUs by hints first.

With multiple devices allocated to a single container, this can
help in cases where the collective locality of devices forces
allocation high in the pool tree, where we should prefer CPUs
with locality to one of the devices and avoid other CPUs.

For instance, devices with locality to NUMA node #0 and #3, or
'half of' sockets #0 and #1, and a request for 2 CPUs, we end
up in the root pool. But we should only prefer allocating CPUs
with locality to NUMA nodes #0 or #3 and avoiding any CPU with
locality to node #1 or #3.

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>