e2e: resize VM / filesystem to max.
Resize e2e test VM root filesystem to the maximum possible space
during provisioning... on Fedora. We don't try to be fancy, only
checking the root filesystem to be btrfs. For the rest we assume
(blindly) a fedora cloud image, where the layout is deliberately
designed to make this near trivial.

On ubuntu this is not necessary as the filesystem ends up being
maxed out automatically, at least on 22.04, and Vagrant support
for 24.04 has been discontinued. We should drop Ubuntu support,
too. Or at least seriously consider it.

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>