e2e: support "cpus-present" in topology.json
cpus-present enables creating virtual machines with different
numbers of possible and enabled CPUs. CPUs that are possible
but not enabled can be hot-plugged when the virtual machine is
running.

q35 machine, irq split and x2apic are required for hot-plugging CPUs
with id >= 256. In case of "separated_output_vars", they are now
printed in place of former "CPU:" parameter. This moves Qemu "-cpu"
parameter management to topology2qemuopts while it used to be
hardcoded "-cpu host" in Vagrantfile.in. Finally, new "SMP:" parameter
printing is added for "-smp" parameters.

Signed-off-by: Antti Kervinen <antti.kervinen@intel.com>