e2e: enable user to override file.var.* with env vars
Currently the test framework overrides an environment variable varname
with a value in varname.var file *if* varname has been created from
varname.var but with larger scope. For instance,
kernel_getsource.var="vanilla" in topology scope that contains many
tests is overridden by kernel_getousrce.var="cxl" file in individual
test scope under the same topology.

This change replaces the *if* with *only if*. That is, if the
environment variable already exists but it has not been created from
varname.var file, it is interpreted as a user choice to run a test
with non-default environment variables, like:

kernel_getsource="git clone /hack/linux" ./run_tests.sh path/to/test

Signed-off-by: Antti Kervinen <antti.kervinen@intel.com>