.github: fix unstable chart publishing with Helm v4.x.
Unstable Helm chart publishing has been broken since Helm 4.x
came out. This is due to our questionable practice to install
and use the latest Helm release in CI, combined with Helm 4.x
refusing to parse a registry URL with a path suffix in login.
Fix the latter of these in our workflow.

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>