e2e: report skipped tests with SKIP instead of PASS in the summary
If a test reported its own verdict to be SKIP, then use that verdict
in the test result summary instead of exit status based verdict from
the framework. If a test provides extra information for the
SKIP (unsupported distro, too old kernel, ...), that will be included
after the test name in the summary.

Signed-off-by: Antti Kervinen <antti.kervinen@intel.com>